refactor: move hardware sheets
