+++ app/src/main/java/to/bitkit/ui/sheets/hardware/HwIntroSheet.kt
@@ -1,4 +1,4 @@
-package to.bitkit.ui.sheets
+package to.bitkit.ui.sheets.hardware
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
