+++ app/src/main/java/to/bitkit/viewmodels/AppViewModel.kt
@@ -145,8 +145,8 @@ import to.bitkit.ui.Routes
 import to.bitkit.ui.components.Sheet
 import to.bitkit.ui.shared.toast.ToastEventBus
 import to.bitkit.ui.shared.toast.ToastQueueManager
-import to.bitkit.ui.sheets.HardwareRoute
 import to.bitkit.ui.sheets.SendRoute
+import to.bitkit.ui.sheets.hardware.HardwareRoute
 import to.bitkit.ui.theme.TRANSITION_SCREEN_MS
 import to.bitkit.usecases.FormatMoneyValue
 import to.bitkit.utils.AppError
